Added notes for file format.
Git applies a thin layer on files when saving -- using a windows client, text is auto converted from CRLF to LF on the backend. This is transparent with git clients (e.g. a checkout will convert LF to CRLF). However, raw.githubusercontent.com does not filter these files, and therefore files obtained via this way have LF (Unix) endings, causing hash validation to fail.

The simple fix here is to ensure that modules are saved with LF endings.

Alternatively (if this does not work), we can set ps and psm modules to be identified as binaires which should force git to ignore any modifications, but we'll lose diff'ing abilities.