Patch issue #4
Fixes issue #4 via the following:

VerifyFile now tests for file existence before calculating the hash of the file. This will prevent 0-length file creation

ValidateAndUpdate: Move New-Item after checking for testing. This means that only 0-length files are created if we know we're going to be able to download the new file from the server.