Refactor classes (issue #6)
Refactored TweekModule and FileManager to configure themselves with command line parameters passed, instead of passing each command line parameter for each method in the class.

A constructor was not used, as these module are dynmically loaded, and therefore, a separate Configure() method was created to modify those values after construction. Think of it as a constructor that doesn't actually create a new object, just sets a bunch of values (e.g. a Setter).